chore: rename IZzzoCharacter.ts to IZzzoAgent.ts
