+++ libs/zzz/db/src/Interfaces/IZzzoAgent.ts
