+++ libs/zzz/db/src/Interfaces/index.ts
@@ -1,4 +1,4 @@
-export * from './IZzzoCharacter'
+export * from './IZzzoAgent'
 export * from './IZzzoDatabase'
 export * from './IZzzoWEngine'
 export * from './IZzzoDriveDisc'
